+++ Changes.md
@@ -14,6 +14,18 @@ Improvements
   - Added visualisation of light filters for USD lights.
   - Added support for USD lights and shaders in the floating inspector panel.
 - ShaderTweaks/ShaderQuery : Added presets for USD light and surface shaders.
+- Test app :
+  - The `-category` argument now accepts a space-separated list of categories, optionally containing wildcards.
+  - Added `-excludedCategories` and `-showCategories` arguments.
+  - Added information about performance test timings to the output stream.
+
+API
+---
+
+- TestRunner :
+  - Added `CategorisedTestMethod` decorator used to assign categories to test methods.
+  - Added `categories()` static method to return the list of available categories in a test suite.
+  - Added `filterCategories()` static method to skip tests according to included and excluded categories.
 
 Fixes
 -----